Redesigning Vm() WIP
TODO: test setC

TODO: and more

WHY: I was trying to test with non-comptime reader and
WHY: writer; I could not do that because my reader and
WHY: writer were not comptime; the Vm() used to take
WHY: three arguments to return a type, which is const
WHY: so it must be calculated during comptime. I got to
WHY: the idea to separate I/O and Dis computer itself.